setting al data types to double to make sure Dtype is always double

+++ tools/caffe_orig.cpp
@@ -105,8 +105,8 @@ int train() {
   }
 
   LOG(INFO) << "Starting Optimization";
-  shared_ptr<caffe::Solver<double> >
-    solver(caffe::GetSolver<double>(solver_param));
+  shared_ptr<caffe::Solver<float> >
+    solver(caffe::GetSolver<float>(solver_param));
 
   if (FLAGS_snapshot.size()) {
     LOG(INFO) << "Resuming from " << FLAGS_snapshot;
@@ -140,24 +140,24 @@ int test() {
   }
   // Instantiate the caffe net.
   Caffe::set_phase(Caffe::TEST);
-  Net<double> caffe_net(FLAGS_model);
+  Net<float> caffe_net(FLAGS_model);
   caffe_net.CopyTrainedLayersFrom(FLAGS_weights);
   LOG(INFO) << "Running for " << FLAGS_iterations << " iterations.";
 
-  vector<Blob<double>* > bottom_vec;
+  vector<Blob<float>* > bottom_vec;
   vector<int> test_score_output_id;
-  vector<double> test_score;
-  double loss = 0;
+  vector<float> test_score;
+  float loss = 0;
   for (int i = 0; i < FLAGS_iterations; ++i) {
-    double iter_loss;
-    const vector<Blob<double>*>& result =
+    float iter_loss;
+    const vector<Blob<float>*>& result =
         caffe_net.Forward(bottom_vec, &iter_loss);
     loss += iter_loss;
     int idx = 0;
     for (int j = 0; j < result.size(); ++j) {
-      const double* result_vec = result[j]->cpu_data();
+      const float* result_vec = result[j]->cpu_data();
       for (int k = 0; k < result[j]->count(); ++k, ++idx) {
-        const double score = result_vec[k];
+        const float score = result_vec[k];
         if (i == 0) {
           test_score.push_back(score);
           test_score_output_id.push_back(j);
@@ -175,10 +175,10 @@ int test() {
   for (int i = 0; i < test_score.size(); ++i) {
     const std::string& output_name = caffe_net.blob_names()[
         caffe_net.output_blob_indices()[test_score_output_id[i]]];
-    const double loss_weight =
+    const float loss_weight =
         caffe_net.blob_loss_weights()[caffe_net.output_blob_indices()[i]];
     std::ostringstream loss_msg_stream;
-    const double mean_score = test_score[i] / FLAGS_iterations;
+    const float mean_score = test_score[i] / FLAGS_iterations;
     if (loss_weight) {
       loss_msg_stream << " (* " << loss_weight
                       << " = " << loss_weight * mean_score << " loss)";
@@ -206,22 +206,22 @@ int time() {
   }
   // Instantiate the caffe net.
   Caffe::set_phase(Caffe::TRAIN);
-  Net<double> caffe_net(FLAGS_model);
+  Net<float> caffe_net(FLAGS_model);
 
   // Do a clean forward and backward pass, so that memory allocation are done
   // and future iterations will be more stable.
   LOG(INFO) << "Performing Forward";
   // Note that for the speed benchmark, we will assume that the network does
   // not take any input blobs.
-  double initial_loss;
-  caffe_net.Forward(vector<Blob<double>*>(), &initial_loss);
+  float initial_loss;
+  caffe_net.Forward(vector<Blob<float>*>(), &initial_loss);
   LOG(INFO) << "Initial loss: " << initial_loss;
   LOG(INFO) << "Performing Backward";
   caffe_net.Backward();
 
-  const vector<shared_ptr<Layer<double> > >& layers = caffe_net.layers();
-  vector<vector<Blob<double>*> >& bottom_vecs = caffe_net.bottom_vecs();
-  vector<vector<Blob<double>*> >& top_vecs = caffe_net.top_vecs();
+  const vector<shared_ptr<Layer<float> > >& layers = caffe_net.layers();
+  vector<vector<Blob<float>*> >& bottom_vecs = caffe_net.bottom_vecs();
+  vector<vector<Blob<float>*> >& top_vecs = caffe_net.top_vecs();
   const vector<vector<bool> >& bottom_need_backward =
       caffe_net.bottom_need_backward();
   LOG(INFO) << "*** Benchmark begins ***";
